unbounded datasets (e.g. with IPv4 subnet, or TLD) are too big. The min-count option helps, but suffers from a couple of problems. (1) it only indirectly accomplishes what we want.  We need a-priori knowledge about the data to set it correctly.  (2) for a 2D dataset, useful data may be lost for less common first-D values.
This patch adds a new feature: the rows (2nd-D) in a column (1st-D)
are now sorted before printing.  [previously they were sorted by
key].  You can specify a limit on the number of rows to print.  Now
a column with low counts may still print a lot of rows.

Data that is not printed due to the max-cells=NNN option are counted
in the SKIPPED and SKIPPED_SUM cells.