qtype_vs_tld and client_subnet2_accum were the same but different.
For whatever reason, we swap the order of the 1st and 2nd dimensions
of the data that comes from the collector.  In one case the swap
happened during extraction (storage) and in the other it happened
during graphing.

Now both are swapped during extraction and there is no need to swap
during graphing.

This was done so that I need only one 'trimmer' function for both
datasets.