Patches for building DSC on Mac OS/X
MacOS Tiger compiler has a bug(?) that caused
TmfBase/Hapy/src/include/Hapy/String.h to be included when
for #include <Hapy/String.h>.  Part of the problem may be that
OS/X treats filenames case-insensitively.

This patch renames the file to HapyString.h.

Other fixes are for DNS #defines