This patch solves the problem of integrating ip_message_callback() and the new IPv6 features.
The former ip_message_callback() accepted a 'struct ip *' as its
argument.  We probably could have gotten away with keeping it that
way if we only wanted to extract the IPv[46] source and destination
addresses from the IP header.  But we also want the payload protocol
type, which is handled differently for v6.  The first solution I
tried turned out to be pretty ugly and required passing a length
arg to all the indexers.

The solution in this patch creates a new structure, ip_message,
which is analogous to the dns_message struct used by all the
DNS indexers.  The downsides to this solution are that
we have some extra copies to convert in_addr and in6_addr to
inX_addr, and we call ip_message_callback() in two places
instead of one.  However, this solution seems to be the least
disruptive overall.

Another downside is that both ip_message and dns_message
have copies of the source/destination IP addresses, although
in dns_message its "client" rather than src/dst.  I considered
maybe having dns_message point to ip_message, but the
way handlers in pcap.c work that would be tricky.