Somewhat significant change.
Ripped out all the "ip_message" and migrated it to use the
"transport_message" structure, which is linked under dns_message.

The separation of IP and DNS processing prevented us from building
a dataset that combined IP and DNS indexers. The transport_message
structure already existed and contained most of what was in ip_message
(only had to add ip_version).

Now we can combine indexers.  This change also means that the "layer"
field of the dataset line in dsc.conf is ignored.  They are all "dns"
indexers now.