bugfix: handle receiving TCP dns length prefix out of order.
A user reported that when a TCP segment containing only the DNS message length
is received *after* the message it references (i.e., out-of-order), then DSC
goes into a 100% CPU loop.  Confirmed that the code doesn't work correctly
when dnslen comes out-of-order, but I wasn't able to easily reproduce the
100% CPU bug.

The fix is to add a "dnslen_bytes_seen_mask" variable that tracks which
of the two dnslen bytes we've seen.  Once both bytes have been seen,
then we can proceed to reassembling the message buffer.