Fix pcap/select last_ts
In some cases `select` will return the fd set as if there are packets
to read but there aren't. That would case the last timestamp to not
advance and the `Pcap_run` loop to never finish.
This fix add a check on packets captured and sets last timestamp if
none where caught.